Arreglos pago de credito y se adicion pago de credito de terceros

+++ lib/modules/home/UserSessionInfo/data/models/user_session_info_model.dart
@@ -93,9 +93,8 @@ class UserInfoResponseModel extends UserInfoResponseEnttity {
         personName: json["personName"],
         idcQuestion: json["idcQuestion"],
         cellPhoneNumber: json["cellPhoneNumber"],
-        maximumAmount: json["maximumAmount"],
-        maximumElectronicWalletAmount: json["maximumElectronicWalletAmount"]
-            .toDouble(),
+        maximumAmount: json["maximumAmount"].toDouble(),
+        maximumElectronicWalletAmount: json["maximumElectronicWalletAmount"].toDouble(),
         email: json["email"],
         sendEmailNotification: json["sendEmailNotification"],
         identityCardNumber: json["identityCardNumber"],
@@ -185,7 +184,7 @@ class ListCodeCreditLineElementModel extends ListCodeCreditLineElementEntity {
       ListCodeCreditLineElementModel(
         idMoney: json["idMoney"],
         codMoney: json["codMoney"],
-        idOperationEntity: json["idOperationEntity"].toDouble(),
+        idOperationEntity: json["idOperationEntity"],
         operationCode: json["operationCode"],
         availableAmount: json["availableAmount"].toDouble(),
         idOffice: json["idOffice"],
