+++ lib/modules/home/UserSessionInfo/domain/entities/user_session_info_entity.dart
@@ -27,7 +27,7 @@ class UserInfoResponseEnttity {
   final dynamic personName;
   final String idcQuestion;
   final dynamic cellPhoneNumber;
-  final int maximumAmount;
+  final double maximumAmount;
   final double maximumElectronicWalletAmount;
   final dynamic email;
   final bool sendEmailNotification;
